fix file names & tests
